Add files via upload
The aim of this project is to predict whether the reviews are positive or negative. I have used the text-blob library to get the polarity of the reviews and then converted that polarity into sentiment columns into numeric columns of 0 and 1. Then used data cleaning on the reviews and applied bag of words, and then trained the data on a machine learning model and tested it with some real-life reviews.